+++ servers/rendering_server.h
@@ -1398,7 +1398,7 @@ public:
 
 	/* EVENT QUEUING */
 
-	virtual void draw(bool p_swap_buffers = true, double frame_step = 0.0) = 0;
+	virtual void draw(bool p_swap_buffers = true, float frame_step = 0.0f) = 0;
 	virtual void sync() = 0;
 	virtual bool has_changed() const = 0;
 	virtual void init() = 0;
